+++ packages/webapp/src/navbar/NavBar.tsx
@@ -11,10 +11,10 @@ import { type Tag } from "../api/models";
 import { EditNavBar } from './EditNavBar';
 import { NavItem } from './NavItem';
 import { ViewNavBar } from './ViewNavBar';
-import { SortMenu } from './SortMenu';
+import { ListNavBar } from './ListNavBar';
 import { SearchInput, SearchButton } from "./SearchInput";
 
-export const DesktopNavBar = ({disableEdit = false, showSort = false, showDialog}) => {
+export const DesktopNavBar = ({disableEdit = false, showList = false, showDialog}) => {
   const viewMode = useEditModeStore(state => state.viewMode);
 
   return (
@@ -29,8 +29,8 @@ export const DesktopNavBar = ({disableEdit = false, showSort = false, showDialog
               { viewMode === ViewMode.EDIT && (
                 <EditNavBar showDialog={showDialog}/>
               )}
-              { showSort && viewMode === ViewMode.VIEW && (
-                <SortMenu />
+              { showList && viewMode === ViewMode.VIEW && (
+                <ListNavBar />
               )}
             </div>
             <div className="flex pr-2 space-x-4">
@@ -43,7 +43,7 @@ export const DesktopNavBar = ({disableEdit = false, showSort = false, showDialog
   )
 }
 
-export const MobileNavBar = ({disableEdit = false, showSort = false, showDialog}) => {
+export const MobileNavBar = ({disableEdit = false, showList = false, showDialog}) => {
   const [showSearch, setShowSearch] = useState(false)
   const viewMode = useEditModeStore(state => state.viewMode);
 
@@ -61,8 +61,8 @@ export const MobileNavBar = ({disableEdit = false, showSort = false, showDialog}
                   { viewMode === ViewMode.EDIT && (
                     <EditNavBar showDialog={showDialog}/>
                   )}
-                  { showSort && viewMode === ViewMode.VIEW && (
-                    <SortMenu />
+                  { showList && viewMode === ViewMode.VIEW && (
+                    <ListNavBar />
                   )}
                 </div>
                 <div className="flex pr-2 space-x-4">
@@ -91,7 +91,7 @@ export const MobileNavBar = ({disableEdit = false, showSort = false, showDialog}
   )
 }
 
-export const NavBar = ({disableEdit = false, showSort = false}) => {
+export const NavBar = ({disableEdit = false, showList = false}) => {
   const [ deviceType ] = useDeviceType();
   const { setDialogVisible, openDialog } = useTagDialog()
 
@@ -113,10 +113,10 @@ export const NavBar = ({disableEdit = false, showSort = false}) => {
   return (
     <>
       { deviceType === DeviceType.DESKTOP &&
-        <DesktopNavBar showDialog={showDialog} disableEdit={disableEdit} showSort={showSort} />
+        <DesktopNavBar showDialog={showDialog} disableEdit={disableEdit} showList={showList} />
       }
       { deviceType === DeviceType.MOBILE &&
-        <MobileNavBar showDialog={showDialog} disableEdit={disableEdit} showSort={showSort} />
+        <MobileNavBar showDialog={showDialog} disableEdit={disableEdit} showList={showList} />
       }
     </>
   )